Moved dto directories to respected modules
