+++ src/tasks/tasks.controller.ts
@@ -9,12 +9,12 @@ import {
   Query,
   UseGuards,
 } from '@nestjs/common';
-import { CreateTaskDto } from 'src/dto/create-task.dto';
-import { TasksService } from './tasks.service';
-import { GetTasksFilterDto } from 'src/dto/get-tasks-filter.dto';
-import { UpdateTaskStatusDto } from 'src/dto/update-task-status.dto';
+import { CreateTaskDto } from 'src/tasks/dto/create-task.dto';
+import { UpdateTaskStatusDto } from 'src/tasks/dto/update-task-status.dto';
 import { Task } from './task.entity';
 import { AuthGuard } from '@nestjs/passport';
+import { TasksService } from './tasks.service';
+import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 
 @Controller('tasks')
 @UseGuards(AuthGuard()) // Guard at the controller level will guard all routes.
